+++ src/main/java/validate/Error.java
@@ -1,4 +1,4 @@
-package racingcar;
+package validate;
 
 /**
  * @author yhh1056
