+++ src/test/java/racingcar/CountTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import validate.Count;
 
 /**
  * @author yhh1056
