+++ src/test/java/racingcar/NamesTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import validate.Names;
 
 /**
  * @author yhh1056
